:recycle:  fix circular dependencies from printer-legacy
